requirements.txt
django-contrib-comments-1.9.1
